% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpal.R
\name{team_colors}
\alias{team_colors}
\title{Sports team color schemes.}
\usage{
team_colors(name)
}
\arguments{
\item{name}{Name of the team to use.}
}
\value{
A vector of colors (hex format)
}
\description{
These are color schemes from a variety of popular teams.
}
\examples{
team_colors("Winnipeg Jets")
}
\keyword{colors}
\keyword{sports}
